Add a variant of cuckoo filters in which buckets may overlap.
This demonstrates lookup times that are as good as cuckoo filter
lookup times, or better, while having space usage close to that of the
semi-sorted cuckoo filter.

This is based on Lehman, Eric, and Rina Panigrahy. "3.5-way cuckoo
hashing for the price of 2-and-a-bit." European Symposium on
Algorithms. Springer, Berlin, Heidelberg, 2009.