Use 8 bits per item as the minimum for SimdBlockFilters.
The optimal value is the number of bits set per Add() divided by ln 2,
which is 8/ln 2 = 11.54. Setting the minimum to 8 bits per item keeps
the number of bits per item in this test always between 8 and 16,
which is closer to the optimal value than the previous range of 12.5
to 25 bits per item.